requirement.txt
beautifulsoup4==4.4.1
chardet==2.3.0
click==6.6
Django==1.10.4
django-crispy-forms==1.6.1
django-registration-redux==1.4
feedparser==5.2.1
Flask==0.11.1
html5lib==0.999
itsdangerous==0.24
Jinja2==2.8
lxml==3.6.4
MarkupSafe==0.23
oauthlib==2.0.1
pymongo==3.3.1
requests==2.12.3
requests-oauthlib==0.7.0
six==1.10.0
tweepy==3.5.0
Werkzeug==0.11.11
